pyproject.toml
# ─── Métadonnées du projet ───────────────────────────────────────────────────
[project]
name = "skillgap-analyzer"
version = "0.1.0"
description = "Data jobs market skill gap analyzer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.1.0",
    "bitsandbytes>=0.49.2",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.4,<3.2",
    "httpx>=0.28.1",
    "huggingface-hub>=1.11.0",
    "llama-cpp-python>=0.3.20",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "pydantic>=2.13.2",
    "tenacity>=9.1.4",
    "torch>=2.11.0",
    "transformers>=5.5.4",
]




# ─── Dépendances de développement (uv) ───────────────────────────────────────
# Séparées des deps runtime pour ne pas les embarquer en prod.
# Gérées par uv dans le venv local, non publiées avec le package.
[dependency-groups]
dev=[
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "pytest-asyncio>=0.21",
  "ruff>=0.1",
  "mypy>=1.5",
  "pre-commit>=4.5.1",
  "pandas-stubs>=3.0.0.260204",
  "dbt-core>=1.11.8",
  "dbt-bigquery>=1.11.1",
]

# ─── Ruff : linter/formatter ──────────────────────────────────────────────────
# Remplace flake8, isort et une partie de pylint en un seul outil.
[tool.ruff]
line-length = 100      # longueur max des lignes (plus permissif que le 88 de black)
target-version = "py312"

# Règles activées :
#   E/W  → style PEP8 (erreurs et warnings)
#   F    → pyflakes (variables inutilisées, imports manquants…)
#   I    → isort (ordre des imports)
#   UP   → pyupgrade (modernisation syntaxe Python)
#   B    → bugbear (patterns risqués ou sous-optimaux)
#   SIM  → simplification de code
#   PTH  → préférer pathlib à os.path
[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "PTH"]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"


# ─── Mypy : vérification statique des types ───────────────────────────────────
# strict = true active toutes les vérifications (no implicit Any, no untyped defs…)
# Garantit une base de code typée de bout en bout.
[tool.mypy]
strict = true
python_version = "3.12"

# ─── Pytest ───────────────────────────────────────────────────────────────────
# testpaths : où pytest cherche les tests (évite de scanner tout le repo)
# addopts   : options appliquées à chaque `pytest` — ici couverture automatique
#             avec affichage des lignes non couvertes dans le terminal
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/skillgap --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillgap"]

[tool.coverage.run]
omit = ["src/skillgap/__main__.py"]

[[tool.mypy.overrides]]
module = ["pyarrow", "pyarrow.*"]
ignore_missing_imports = true
